+++ src/steps/Microsoft.DotNet.UpgradeAssistant.Steps.Razor/ITextReplacer.cs
@@ -1,0 +1,12 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Collections.Generic;
+
+namespace Microsoft.DotNet.UpgradeAssistant.Steps.Razor
+{
+    public interface ITextReplacer
+    {
+        void ApplyTextReplacements(IList<TextReplacement> replacements);
+    }
+}
