+++ tests/tool/Integration.Tests/IntegrationScenarios/AspNetSample/csharp/Original/Views/Home/Index.cshtml
