+++ tests/tool/Integration.Tests/IntegrationScenarios/AspNetSample/csharp/Upgraded/Views/Home/Index.cshtml
